refactored db reset to utils.js
