+++ server/test/post_tests.js
@@ -6,32 +6,11 @@ chai.use(chaiHttp);
 const POST_USER_CASES = require('./test_cases/register_user.json');
 const POST_POST_CASES = require('./test_cases/create_post.json');
 const POST_COMMENT_CASES = require('./test_cases/add_comment.json');
-const Board = require("../models/board");
-const User = require("../models/user");
-const Post = require("../models/post");
-const MODEL_DATA = require("./test_cases/model_data.json");
-
+const utils = require('./utils');
 
 describe('/POST tests', () => {
     beforeEach(() => {
-        Board.deleteMany({});
-        User.deleteMany({});
-        Post.deleteMany({});
-        MODEL_DATA.boards.forEach((model) => {
-            const newBoard = new Board(model);
-            newBoard.save();
-        });
-        MODEL_DATA.users.forEach((model) => {
-            const newUser = new User(model);
-            newUser.save();
-        });
-        MODEL_DATA.posts.forEach((model) => {
-            const newPost = new Post(model);
-            newPost.save()
-                .then((post) => {
-                    console.log("model post:" + post)
-                });
-        });
+        utils.resetDB();
     });
     describe('User registration', () => {
         POST_USER_CASES.forEach(doPost);
